+++ NEWS.md
@@ -1,8 +1,8 @@
 # araponga 1.1.0
 
 * `download.simdata()`: regenerated simulation dataset v1.1.0, now with fixed floating-point issue.
-* `download.simdata()`: updated to download dataset v1.1.0 and warn users that might have the older version installed.
-* `download.simdata()`: added codes for generating dataset to `data-raw/simdata`
+* `download.simdata()` & `find.3d()`: updated to download dataset v1.1.0 and warn users that might have the older version installed.
+* `download.simdata()`: added codes for generating dataset to `data-raw/simdata`.
 * `pitch2d.from.3d()`: fixed floating-point handling for degenerate projections (#4).
 * `find.3d()`: fixed bug that dropped column name in the output.
 * `find.3d()` and others: minor clarification in documentation for `label_error` argument.